Modular Effects spend energy + improved effect selection screen

+++ modular-armor/src/main/kotlin/dev/nathanpb/dml/modular_armor/effects/AbilityBasedEffect.kt
@@ -34,7 +34,7 @@ import net.minecraft.util.Identifier
 abstract class AbilityBasedEffect(
     id: Identifier,
     category: EntityCategory,
-    applyCost: () -> Float,
+    applyCost: () -> Long,
     val ability: PlayerAbility
 ) : ModularEffect<ModularEffectTriggerPayload>(id, category, applyCost) {
