+++ modular-armor/src/main/kotlin/dev/nathanpb/dml/modular_armor/effects/DamageImmunityLikeEffect.kt
@@ -33,7 +33,7 @@ import net.minecraft.util.Identifier
 abstract class DamageImmunityLikeEffect(
     id: Identifier,
     category: EntityCategory,
-    applyCost: ()->Float
+    applyCost: ()->Long
 ) : ProtectionLikeEffect(id, category, applyCost) {
 
     override fun registerEvents() {
